Fixes issue #2282 - Refactor multipart support
