--- extension/apache-fileupload/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,0 @@
-cloud.piranha.extension.apache.fileupload.ApacheMultiPartInitializer
