+++ extension/standard/pom.xml
@@ -20,6 +20,12 @@
         <!-- compile -->
         <dependency>
             <groupId>cloud.piranha.extension</groupId>
+            <artifactId>piranha-extension-apache-fileupload</artifactId>
+            <version>${project.version}</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>cloud.piranha.extension</groupId>
             <artifactId>piranha-extension-annotationscan</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
@@ -60,11 +66,11 @@
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
-        
         <dependency>
             <groupId>cloud.piranha.extension</groupId>
             <artifactId>piranha-extension-security-servlet</artifactId>
             <version>${project.version}</version>
+            <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>cloud.piranha.extension</groupId>
@@ -98,13 +104,6 @@
             <scope>provided</scope>
         </dependency>
         <!-- runtime -->
-        <dependency>
-            <groupId>cloud.piranha.extension</groupId>
-            <artifactId>piranha-extension-apache-fileupload</artifactId>
-            <version>${project.version}</version>
-            <classifier>shaded</classifier>
-            <scope>runtime</scope>
-        </dependency>
         <dependency>
             <groupId>jakarta.annotation</groupId>
             <artifactId>jakarta.annotation-api</artifactId>
