+++ extension/standard/src/main/java/module-info.java
@@ -48,6 +48,7 @@
 module cloud.piranha.extension.standard {
     exports cloud.piranha.extension.standard;
     opens cloud.piranha.extension.standard;
+    requires cloud.piranha.extension.apache.fileupload;
     requires cloud.piranha.extension.annotationscan;
     requires cloud.piranha.extension.herring;
     requires cloud.piranha.extension.locale_encoding;
